Give the ability in freestyle jobs to update PagerDuty change events
This allows Jenkins freestyle jobs to update change events using the
same fields as the PagerDuty post build step, except integration token
and summary text are available.

In this way, a freestyle job might use this build step in the
beginning of the build to mark a change event before any other regular
Jenkins build steps. (shell block, etc) Another PagerDuty change event
can be addded at the end to indicate that the job is done.